pyproject.toml
[project]
name = "electricity-pricing-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.5",
    "openelectricity>=0.9.2",
    "pandas>=3.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
]
